Dala 1270 remove unreachable pages (#187)
* delete the 3 pages to be removed

* adapt tests

* migrate test for alternative name search

---------
