+++ dataland-frontend/tests/e2e/specs/framework-view-and-upload-journey/RoleChecks.ts
@@ -1,9 +1,9 @@
-import { reader_name, reader_pw, uploader_name, uploader_pw } from "../../utils/Cypress";
-import { DataTypeEnum } from "../../../../build/clients/backend";
-import { describeIf } from "../../support/TestUtility";
-import { uploadCompanyViaApi } from "../../utils/CompanyUpload";
-import { getKeycloakToken } from "../../utils/Auth";
-import { generateCompanyInformation } from "../../fixtures/CompanyFixtures";
+import { reader_name, reader_pw, uploader_name, uploader_pw } from "@e2e/utils/Cypress";
+import { DataTypeEnum } from "@clients/backend";
+import { describeIf } from "@e2e/support/TestUtility";
+import { uploadCompanyViaApi } from "@e2e/utils/CompanyUpload";
+import { getKeycloakToken } from "@e2e/utils/Auth";
+import { generateCompanyInformation } from "@e2e/fixtures/CompanyFixtures";
 
 describe("Check if each page is visitable if and only if the corresponding role is given", () => {
   describeIf(
@@ -25,19 +25,16 @@ describe("Check if each page is visitable if and only if the corresponding role 
               `/samples/${DataTypeEnum.EutaxonomyNonFinancials}`,
               "/companies",
               "/datasets",
-              "/companies-only-search",
               "/requests",
               "/api-key",
               "/dataprivacy",
               "/imprint",
               "/nocontent",
-              `/companies/${company.companyId}`,
               `/companies/${company.companyId}/frameworks/${DataTypeEnum.EutaxonomyFinancials}`,
               `/companies/${company.companyId}/frameworks/${DataTypeEnum.EutaxonomyNonFinancials}`,
               `/companies/${company.companyId}/frameworks/${DataTypeEnum.Lksg}`,
             ];
             uploaderPages = [
-              "/companies/upload",
               "/companies/choose",
               `/companies/${company.companyId}/frameworks/upload`,
               `/companies/${company.companyId}/frameworks/${DataTypeEnum.EutaxonomyFinancials}/upload`,
